gst: src: Improve location handling
Move all URI handler code into source element. It is meant for src only and
not to be shared with other elements. This allows to merge set_location function
into a single one that better guards against unsupported URIs.

Also replace gtuber scheme when passing URI to Gtuber for media info fetching,
as gtuber API does not guarantee support for it.