gtuber: Drop libsoup2 support
It was hacky and problematic as not only plugins needed to support both versions, but gtuber
implementations also had to be compiled against the same libsoup version as gtuber itself.

This is the right thing to do as newly written libs should target libsoup3 and it enables
HTTP/2 support which allows us to handle more websites. GStreamer since 1.19.90 works with
libsoup3, thus the minimal version bump.